+++ NEWS.md
@@ -1,5 +1,10 @@
 # NEWS for vip package
 
+### Changes for version 0.0.2
+
+* Added support for XGBoost models (i.e., objects of class `"xgb.booster"`).
+* Added support for ranger models (i.e., objects of class `"ranger"`).
+
 
 ### Changes for version 0.0.1